+++ src/appData/personal.tsx
@@ -1,10 +1,3 @@
-import { Codepen, Facebook, GithubIcon, Instagram, LinkedIn, X } from '@/utils/icons'
+import { getSocialLinks } from '@/services/site'
 
-export const socials = [
-  { name: 'Github', href: 'https://github.com/Marlon042/', icon: <GithubIcon /> },
-  { name: 'LinkedIn', href: 'https://www.linkedin.com/in/marlon-gutierrez-v/', icon: <LinkedIn /> },
-  { name: 'Codepen', href: '#', icon: <Codepen /> },
-  { name: 'X', href: '#', icon: <X /> },
-  { name: 'Instagram', href: '#', icon: <Instagram /> },
-  { name: 'Facebook', href: '#', icon: <Facebook /> },
-]
+export { getSocialLinks }